MSP430 Hall Effect Sensor Code
Should be complete. Calculates speed in MPH and stores as a UQ6.2
fixed-point number. Uses a single-byte SPI transfer (interrupts to be
removed).

Changes to original code:
- Tightened up the detection of the pulses from the ADC values.
Increased from 450 - 700.
- Removed RPM calculation function to mitigate overhead from function
calls.
- Includes a 0.8 second time-out - Any speed below ~5 MPH will be read
as 0 MPH. This prevents the MSP from transmitting old speed data if a
hang-up occurs.

To be done:
- Remove SPI interrupts (test to see whether or not that actually works)
- Change threshold for detection based on the distance between magnet
and the sensor
- Test to see if it actually works